Complete consolidatie van de tools gedaan en het tools bestand van singleagent hernoemd alsook de referenties ernaar
