+++ .gitignore
@@ -1,4 +1,5 @@
 /venv
+/.venv
 *.DS_Store
 __pycache__/
 *.pyc
